Refactored nutrition journal page
